backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.0
scikit-learn==1.3.0
beautifulsoup4==4.12.3
requests==2.31.0
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
faiss-cpu==1.7.4
numpy==1.26.4
sentence-transformers==2.5.1
huggingface_hub==0.20.3
gunicorn==21.2.0